refactor to traits and apply namer
Parsing and TagGeneration are now traits that are mixed into SCTags so that
they can have direct access to the compiler instead of messing around with the
implicit arguments. This also helps alot with the types that have to be
written down.

The namer stage is also run on the code now so that symbols in the AST are
populated. This will help with adding additional info to the tags file.

darcs-hash:20080120205424-8d81a-0a6519ab892952c89d1d7efc85858e61abc1b245.gz